requirements.txt
tornado
mysql-connector-python
sqlalchemy
sockjs-tornado
wtforms
werkzeug
email_validator